creating general_mcp
